requirements.txt
django==3.2.12
djangorestframework==3.13.1
django-filter==21.1
requests==2.27.1